feat: rename variant Effect
