+++ apps/platform/src/pages/VariantPage/Profile.tsx
@@ -9,7 +9,7 @@ import {
 } from "ui";
 
 import PharmacogenomicsSummary from "sections/src/variant/Pharmacogenomics/Summary";
-import InSilicoPredictorsSummary from "sections/src/variant/InSilicoPredictors/Summary";
+import VariantEffectSummary from "sections/src/variant/VariantEffect/Summary";
 import VariantEffectPredictorSummary from "sections/src/variant/VariantEffectPredictor/Summary";
 import EVASummary from "sections/src/variant/EVA/Summary";
 import UniProtVariantsSummary from "sections/src/variant/UniProtVariants/Summary";
@@ -18,9 +18,7 @@ import QTLCredibleSetsSummary from "sections/src/variant/QTLCredibleSets/Summary
 
 import ProfileHeader from "./ProfileHeader";
 const PharmacogenomicsSection = lazy(() => import("sections/src/variant/Pharmacogenomics/Body"));
-const InSilicoPredictorsSection = lazy(
-  () => import("sections/src/variant/InSilicoPredictors/Body")
-);
+const VariantEffectSection = lazy(() => import("sections/src/variant/VariantEffect/Body"));
 const VariantEffectPredictorSection = lazy(
   () => import("sections/src/variant/VariantEffectPredictor/Body")
 );
@@ -31,7 +29,7 @@ const QTLCredibleSetsSection = lazy(() => import("sections/src/variant/QTLCredib
 
 const summaries = [
   PharmacogenomicsSummary,
-  InSilicoPredictorsSummary,
+  VariantEffectSummary,
   VariantEffectPredictorSummary,
   EVASummary,
   UniProtVariantsSummary,
@@ -67,7 +65,7 @@ function Profile({ varId }: ProfileProps) {
       <ProfileHeader />
 
       <SummaryContainer>
-        <InSilicoPredictorsSummary />
+        <VariantEffectSummary />
         <VariantEffectPredictorSummary />
         <EVASummary />
         <UniProtVariantsSummary />
@@ -78,7 +76,7 @@ function Profile({ varId }: ProfileProps) {
 
       <SectionContainer>
         <Suspense fallback={<SectionLoader />}>
-          <InSilicoPredictorsSection id={varId} entity={VARIANT} />
+          <VariantEffectSection id={varId} entity={VARIANT} />
         </Suspense>
         <Suspense fallback={<SectionLoader />}>
           <VariantEffectPredictorSection id={varId} entity={VARIANT} />
