--- packages/sections/src/variant/InSilicoPredictors/InSilicoPredictorsSummaryFragment.gql
@@ -1,5 +1,0 @@
-fragment InSilicoPredictorsSummaryFragment on Variant {
-  inSilicoPredictors {
-    method
-  }
-}
