+++ packages/sections/src/variant/VariantEffect/Description.tsx
